# target/swagger.json
{"swagger":"2.0","info":{"version":"1.7.0","title":"D-StreaMon companion app REST API","description":"D-StreaMon companion app REST API"},"basePath":"/onos/d-streamon","produces":["application/json"],"consumes":["application/json"],"tags":[{"name":"mgmts","description":"Web resource for retrieving MgmtInterface objects"},{"name":"stacks","description":"Web resource for storing Stack objects"}],"paths":{"/mgmts/{probeUuid}":{"get":{"tags":["mgmts"],"summary":"Retrieves the name of the mgmt port","description":"","parameters":[{"name":"probeUuid","in":"path","type":"string","description":"the associated probe uuid","required":true}],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/mgmts":{"get":{"tags":["mgmts"],"summary":"Returns the uuids of the probes","description":"","parameters":[],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/stacks":{"post":{"tags":["stacks"],"summary":"Registers the data of a new stack created through OpenStack Heat","description":"","parameters":[{"name":"stream","in":"body","description":"the input stream","required":true}],"consumes":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}},"get":{"tags":["stacks"],"summary":"Returns the uuids of the registered stacks","description":"","parameters":[],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/stacks/{stackUuid}":{"delete":{"tags":["stacks"],"summary":"Uregisters a stack  previously created","description":"","parameters":[{"name":"stackUuid","in":"path","type":"string","description":"the stack uuid","required":true}],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}},"get":{"tags":["stacks"],"summary":"Retrieves the data of a single stack","description":"","parameters":[{"name":"stackUuid","in":"path","type":"string","description":"the uuid of the stack to retrieve","required":true}],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}}},"definitions":{}}